Move validation logic to prevent dialog opening for invalid URLs
