+++ src/components/modebar/buttons.js
@@ -73,9 +73,25 @@ modeBarButtons.sendChartToCloud = {
     title: function(gd) { return _(gd, 'Share with Plotly Cloud'); },
     icon: Icons.cloudupload,
     click: function(gd) {
-        var serverUrl = (window.PLOTLYENV || {}).BASE_URL || gd._context.plotlyServerURL;
+        var baseUrl = (window.PLOTLYENV || {}).BASE_URL || gd._context.plotlyServerURL;
+        if(!baseUrl) {
+            console.error('No destination URL provided (plotlyServerURL is not set)');
+            return;
+        }
+
+        // Plotly Cloud origin, used to validate incoming messages and to target outgoing ones.
+        // `baseUrl` (plotlyServerURL) is the upload page that handles login and signals
+        // back when authentication succeeds.
+        var cloudOrigin;
+        try {
+            cloudOrigin = new URL(baseUrl).origin;
+        } catch(e) {
+            console.error('Invalid plotlyServerURL: ' + baseUrl);
+            return;
+        }
+
         confirmCloudDialog(gd, serverUrl, function() {
-            Plots.sendDataToCloud(gd);
+            Plots.sendDataToCloud(gd, cloudOrigin);
         });
     }
 };
